requirements.txt
bs4
Jinja2
matplotlib
nltk
numpy
opencv-python
pillow
selenium
